+++ packages/terra-framework-docs/src/terra-dev-site/test/folder-tree/BasicFolderTree.test.jsx
@@ -49,21 +49,21 @@ const BasicFolderTree = () => {
           icon={<IconDocuments />}
           key="info"
           isSelected={selectedKey === 'info'}
-          onClick={() => { setSelectedKey('info'); }}
+          onSelect={() => { setSelectedKey('info'); }}
         />
         <FolderTree.Item
           label="projects_2023.txt"
           icon={<IconDocuments />}
           key="projects_2023"
           isSelected={selectedKey === 'projects_2023'}
-          onClick={() => { setSelectedKey('projects_2023'); }}
+          onSelect={() => { setSelectedKey('projects_2023'); }}
         />
         <FolderTree.Item
           label="Projects"
           key="projects"
           isSelected={selectedKey === 'projects'}
           isExpanded={expandedItems.projects}
-          onClick={() => { setSelectedKey('projects'); }}
+          onSelect={() => { setSelectedKey('projects'); }}
           onToggle={() => { handleExpandCollapseKeys('projects'); }}
           subfolderItems={[
             (<FolderTree.Item
@@ -71,21 +71,21 @@ const BasicFolderTree = () => {
               icon={<IconDocuments />}
               key="project_data1"
               isSelected={selectedKey === 'project_data1'}
-              onClick={() => { setSelectedKey('project_data1'); }}
+              onSelect={() => { setSelectedKey('project_data1'); }}
             />),
             (<FolderTree.Item
               label="project_data2.txt"
               icon={<IconDocuments />}
               key="project_data2"
               isSelected={selectedKey === 'project_data2'}
-              onClick={() => { setSelectedKey('project_data2'); }}
+              onSelect={() => { setSelectedKey('project_data2'); }}
             />),
             (<FolderTree.Item
               label="Tests"
               key="tests"
               isSelected={selectedKey === 'tests'}
               isExpanded={expandedItems.tests}
-              onClick={() => { setSelectedKey('tests'); }}
+              onSelect={() => { setSelectedKey('tests'); }}
               onToggle={() => { handleExpandCollapseKeys('tests'); }}
               icon={<IconDocuments />}
               subfolderItems={[
@@ -94,21 +94,21 @@ const BasicFolderTree = () => {
                   icon={<IconDocuments />}
                   key="long-name-test"
                   isSelected={selectedKey === 'long-name-test'}
-                  onClick={() => { setSelectedKey('long-name-test'); }}
+                  onSelect={() => { setSelectedKey('long-name-test'); }}
                 />),
                 (<FolderTree.Item
                   label="test.txt"
                   icon={<IconDocuments />}
                   key="test"
                   isSelected={selectedKey === 'test'}
-                  onClick={() => { setSelectedKey('test'); }}
+                  onSelect={() => { setSelectedKey('test'); }}
                 />),
                 (<FolderTree.Item
                   label="even_longer_long_long_long_long_long_long_long_long_long_long_long_long_name_test.txt"
                   icon={<IconDocuments />}
                   key="longer-name-test"
                   isSelected={selectedKey === 'longer-name-test'}
-                  onClick={() => { setSelectedKey('longer-name-test'); }}
+                  onSelect={() => { setSelectedKey('longer-name-test'); }}
                 />),
               ]}
             />),