+++ packages/terra-folder-tree/CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
+* Breaking Changes
+  * Renamed `onClick` prop to `onSelect`.
+
+* Added
+  * Added `isSelectable` prop to FolderTreeItem.
+
 ## 1.0.0-alpha.8 - (February 16, 2024)
 
 * Added
-  * Added translations
+  * Added translations.
 
 ## 1.0.0-alpha.7 - (February 7, 2024)