+++ packages/terra-framework-docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Changed
   * Updated the `terra-date-time-picker` timezone example.
+  * Updated selection examples and tests for `terra-folder-tree`.
 
 ## 1.66.0 - (February 20, 2024)
 
@@ -19,6 +20,7 @@
 
 * Added
   * Added shouldTrapFocus to be set to 'true'.
+
 * Changed
   * Updated `terra-navigation-side-menu` example to use more meaningful labels.
   * Updated the `terra-date-time-picker` example for field label.